Enable anisotropic filtering for 3D materials by default on desktop
This results in better appearance for textures viewed from oblique
angles. Many other engines now enable it by default.
On today's desktop platforms, the performance impact is minimal if
other effects are already taxing the GPU.

The default value chosen (8x) is a good balance between performance
and quality. Depending on the scene, you may be able to use 16x
anisotropic filtering with little to no cost in terms of frame time.

The anisotropic filtering quality can be configured as usual in the
Project Settings. Setting it to 1x will disable it even if individual
materials have it enabled, so this is easy to revert if you don't
want it.