Refactor BlendShapeBake::convert_scene to use TypedArray for finding AnimationPlayer and MeshInstance3D nodes
The code changes the way AnimationPlayer and MeshInstance3D nodes are found in the scene. Instead of using a queue, it now uses TypedArray to find all AnimationPlayer and MeshInstance3D nodes.

Additionally, the code modifies how animation libraries are added to the AnimationPlayer. The library name is now appended with "_baked" before being added to the AnimationPlayer.